+++ changelog.md
@@ -22,6 +22,7 @@ Codion Change Log
 - EntityType.define() and AttributeDefiner.define() renamed as().
 ### is.codion.framework.db.local
 - DefaultEntityQueries.update() bug fixed, buggy modified check removed, caller is responsible for providing a modified entity.
+- EntityResultPacker, minor refactoring, transient attributes are never derived.
 ### is.codion.swing.framework.model
 - EntityComboBox.Builder.select() added.
 - EntityComboBoxModel.Builder.filter() added.